// app/renderer/src/main/src/services/axios.d.ts
export type AxiosRequestHeaders = Record<string, string | number | boolean>

export type AxiosResponseHeaders = Record<string, string> & {
  'set-cookie'?: string[]
}

export interface AxiosRequestTransformer {
  (data: any, headers?: AxiosRequestHeaders): any
}

export interface AxiosResponseTransformer {
  (data: any, headers?: AxiosResponseHeaders): any
}

export interface AxiosAdapter {
  (config: AxiosRequestConfig): AxiosPromise
}

export interface AxiosBasicCredentials {
  username: string
  password: string
}

export interface AxiosProxyConfig {
  host: string
  port: number
  auth?: {
    username: string
    password: string
  }
  protocol?: string
}

export type Method =
  | 'get'
  | 'GET'
  | 'delete'
  | 'DELETE'
  | 'head'
  | 'HEAD'
  | 'options'
  | 'OPTIONS'
  | 'post'
  | 'POST'
  | 'put'
  | 'PUT'
  | 'patch'
  | 'PATCH'
  | 'purge'
  | 'PURGE'
  | 'link'
  | 'LINK'
  | 'unlink'
  | 'UNLINK'

export type ResponseType = 'arraybuffer' | 'blob' | 'document' | 'json' | 'text' | 'stream'

export type responseEncoding =
  | 'ascii'
  | 'ASCII'
  | 'ansi'
  | 'ANSI'
  | 'binary'
  | 'BINARY'
  | 'base64'
  | 'BASE64'
  | 'base64url'
  | 'BASE64URL'
  | 'hex'
  | 'HEX'
  | 'latin1'
  | 'LATIN1'
  | 'ucs-2'
  | 'UCS-2'
  | 'ucs2'
  | 'UCS2'
  | 'utf-8'
  | 'UTF-8'
  | 'utf8'
  | 'UTF8'
  | 'utf16le'
  | 'UTF16LE'

export interface TransitionalOptions {
  silentJSONParsing?: boolean
  forcedJSONParsing?: boolean
  clarifyTimeoutError?: boolean
}

export interface AxiosRequestConfig<D = any> {
  url?: string
  method?: Method
  baseURL?: string
  transformRequest?: AxiosRequestTransformer | AxiosRequestTransformer[]
  transformResponse?: AxiosResponseTransformer | AxiosResponseTransformer[]
  headers?: AxiosRequestHeaders
  params?: any
  paramsSerializer?: (params: any) => string
  data?: D
  timeout?: number
  timeoutErrorMessage?: string
  withCredentials?: boolean
  adapter?: AxiosAdapter
  auth?: AxiosBasicCredentials
  responseType?: ResponseType
  responseEncoding?: responseEncoding | string
  xsrfCookieName?: string
  xsrfHeaderName?: string
  onUploadProgress?: (progressEvent: any) => void
  onDownloadProgress?: (progressEvent: any) => void
  maxContentLength?: number
  validateStatus?: ((status: number) => boolean) | null
  maxBodyLength?: number
  maxRedirects?: number
  socketPath?: string | null
  httpAgent?: any
  httpsAgent?: any
  proxy?: AxiosProxyConfig | false
  cancelToken?: CancelToken
  decompress?: boolean
  transitional?: TransitionalOptions
  signal?: AbortSignal
  insecureHTTPParser?: boolean
}
// ...
export interface AxiosResponse<T = any, D = any> {
  data: T
  status: number
  statusText: string
  headers: AxiosResponseHeaders
  config: AxiosRequestConfig<D>
  request?: any
}

export interface AxiosError<T = any, D = any> extends Error {
  config: AxiosRequestConfig<D>
  code?: string
  request?: any
  response?: AxiosResponse<T, D>
  isAxiosError: boolean
  toJSON: () => object
}

export interface AxiosPromise<T = any> extends Promise<AxiosResponse<T>> {}

export interface CancelStatic {
  new (message?: string): Cancel
}

export interface Cancel {
  message: string | undefined
}

export interface Canceler {
  (message?: string): void
}

export interface CancelTokenStatic {
  new (executor: (cancel: Canceler) => void): CancelToken
  source(): CancelTokenSource
}

export interface CancelToken {
  promise: Promise<Cancel>
  reason?: Cancel
  throwIfRequested(): void
}